Добавляет рефакторинг кода по замечаниям
