+++ src/view/event.js
@@ -1,20 +1,5 @@
 import dayjs from 'dayjs';
-import durationPlugin from 'dayjs/plugin/duration';
-
-dayjs.extend(durationPlugin);
-
-const getDateTimeFormat = (date, isTime) => {
-  return isTime ? dayjs(date).format(`YYYY-MM-DD[T]HH:mm`) : dayjs(date).format(`YYYY-MM-DD`);
-};
-
-const getEventDuration = (dateStart, dateEnd) => {
-  const diff = dayjs(dateEnd).diff(dayjs(dateStart));
-  const days = dayjs.duration(diff).days().toString();
-  const hours = dayjs.duration(diff).hours().toString();
-  const minutes = dayjs.duration(diff).minutes().toString();
-
-  return `${days > 0 ? days + `D` : ``} ${hours > 0 ? hours.padStart(2, `0`) + `H` : ``} ${minutes.padStart(2, `0`)}M`;
-};
+import {getDateTimeFormat, getEventDuration} from '../util';
 
 const renderOffers = (offers) => {
   return `<h4 class="visually-hidden">Offers:</h4>
