+++ src/view/trip-cost.js
@@ -1,7 +1,5 @@
-import {getRandomInteger} from './../util';
-
-export const createTripCostTemplate = () => {
-  const price = getRandomInteger(100, 1000);
+export const createTripCostTemplate = (events) => {
+  const price = events.reduce((acc, event) => acc + event.price, 0);
   return `<p class="trip-info__cost">
     Total: &euro;&nbsp;<span class="trip-info__cost-value">${price}</span>
   </p>`;
